[#664] package CurrenciesView
